Separate test for power_of_two
